We want to find biomarkers that can improve prognostic
We used the Cox model to analyze survival data because it is suitable for handling right-censored survival data and can simultaneously consider the impact of multiple variables on survival time.

In addition to using the Cox proportional hazards model, we also used an exponential survival model to compare the survival probabilities of the high-CXCL12 and low-CXCL12 groups. Exponential survival analysis provides a simple and intuitive method to understand the impact of CXCL12 gene expression levels on survival time